Refactor child workflow replay helpers
