+++ crates/harness-server/src/workflow_runtime_worker.rs
@@ -1,6 +1,7 @@
 mod activity_contract;
 mod activity_result;
 mod child_workflow;
+mod child_workflow_replay;
 mod data_helpers;
 mod executor;
 mod pr_feedback_inspection;
